Factor out getIssuesHandler
